feat: refactor structure and remove unused code
