--- core/constants.js
@@ -1,67 +1,0 @@
-export const EVENT_TYPE_LIST = [
-    'animationcancel',
-    'animationend',
-    'animationiteration',
-    'animationstart',
-    'afterscriptexecute',
-    'auxclick',
-    'beforescriptexecute',
-    'blur',
-    'click',
-    'compositionend',
-    'compositionstart',
-    'compositionupdate',
-    'contextmenu',
-    'copy',
-    'cut',
-    'dblclick',
-    'DOMActivate',
-    'DOMMouseScroll',
-    'error',
-    'focusin',
-    'focusout',
-    'focus',
-    'fullscreenchange',
-    'fullscreenerror',
-    'gesturechange',
-    'gestureend',
-    'gesturestart',
-    'gotpointercapture',
-    'keydown',
-    'keypress',
-    'keyup',
-    'lostpointercapture',
-    'mousedown',
-    'mouseenter',
-    'mouseleave',
-    'mousemove',
-    'mouseout',
-    'mouseover',
-    'mouseup',
-    'mousewheel',
-    'paste',
-    'pointercancel',
-    'pointerdown',
-    'pointerenter',
-    'pointerleave',
-    'pointermove',
-    'pointerout',
-    'pointerover',
-    'pointerup',
-    'scroll',
-    'select',
-    'touchcancel',
-    'touchend',
-    'touchmove',
-    'touchstart',
-    'transitioncancel',
-    'transitionend',
-    'transitionrun',
-    'transitionstart',
-    'webkitmouseforcechanged',
-    'webkitmouseforcedown',
-    'webkitmouseforceup',
-    'webkitmouseforcewillbegin',
-    'wheel',
-    'cancel'
-  ]
